add Gewichtung from Kompetenzbereich

+++ NoVe/Models/FachKompetenzbereich.cs
@@ -16,7 +16,7 @@ namespace NoVe.Models
         public int InEditTime { get; set; }
         public Kompetenzbereich Kompetenzbereich { get; set; }
         public List<NoteView> NoteView { get; set; }
-        public float NotenwertKompetenzbereich { get; set; }
-        public float NotenwertGesamt { get; set; }
+        public double NotenwertKompetenzbereich { get; set; }
+        public double NotenwertGesamt { get; set; }
     }
 }
